Reorganize files and fix tests
