+++ astropy/wcs/wcsapi/__init__.py
@@ -1,3 +1,3 @@
 from .low_level_api import *
-from .base_high_level_api import *
+from .high_level_api import *
 from .high_level_wcs_wrapper import *
